compoundInterest/requirements.txt
mcp>=0.1.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
Pillow>=10.0.0
pyautogui>=0.9.54 